removed unecessary files and renamed authenticaton to router.js
