+++ backend/app.js
@@ -1,6 +1,5 @@
 var express = require('express');
-var path = require('path');
-var favicon = require('serve-favicon');
+// var path = require('path');
 var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
@@ -11,25 +10,17 @@ var app = express();
 var port = require('./SECRET/config').port || 8080;
 require('./app/models/db.js');
 
-
-// view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'jade');
-
-// uncomment after placing your favicon in /public
-//app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
 app.use(cors());
 app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
 
 require('./app/routes/microsoft')(app);
 
-var authentication = require('./app/routes/authentication');
+var router = require('./app/routes/router');
 
-app.use('/api/v1/', authentication);
+app.use('/api/v1/', router);
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
   var err = new Error('Not Found');
@@ -62,6 +53,4 @@ app.use(function(err, req, res, next) {
 });
 
 app.listen(port);
-
-console.log('App listening on localhost:'+port);
 module.exports = app;
