+++ backend/package.json
@@ -14,7 +14,6 @@
     "jade": "~1.11.0",
     "jsonwebtoken": "^7.1.9",
     "mongoose": "^4.6.4",
-    "morgan": "~1.7.0",
-    "serve-favicon": "~2.3.0"
+    "morgan": "~1.7.0"
   }
 }
